MGDCTRS-1418 Add description texts to Debezium ConnectorTypes
closes https://issues.redhat.com/browse/MGDCTRS-1418
