--- src/main/resources/channels-mongodb-2.0.json
@@ -1,16 +1,0 @@
-{
-  "stable": {
-    "shard_metadata": {
-      "connector_revision": 2,
-      "operators": [
-        {
-          "type": "debezium-connector-operator",
-          "version": "[1.0.0,2.0.0)"
-        }
-      ],
-      "connector_type": "source",
-      "connector_class": "io.debezium.connector.mongodb.MongoDbConnector",
-      "container_image": "quay.io/rhoas/cos-connector-debezium-mongodb@sha256:343bb1f834d1a7424e1ff52f773bce77b1d0672724f005546d76c5b202e5b333"
-    }
-  }
-}
