--- src/main/resources/channels-postgres-2.0.json
@@ -1,16 +1,0 @@
-{
-  "stable": {
-    "shard_metadata": {
-      "connector_revision": 2,
-      "operators": [
-        {
-          "type": "debezium-connector-operator",
-          "version": "[1.0.0,2.0.0)"
-        }
-      ],
-      "connector_type": "source",
-      "connector_class": "io.debezium.connector.postgresql.PostgresConnector",
-      "container_image": "quay.io/rhoas/cos-connector-debezium-postgres@sha256:81b1354c9a52b604715507952929a794f46d391807437bc0bd4d31eaeeba9d2b"
-    }
-  }
-}
