+++ src/main/resources/overrides-postgres-2.0.json
@@ -1,0 +1,21 @@
+{
+  "connector_type" : {
+    "description" : "Captures row-level changes in the schemas of a PostgreSQL database by using the `pgoutput` plugin."
+  },
+  "channels": {
+    "stable": {
+      "shard_metadata": {
+        "connector_revision": 2,
+        "operators": [
+          {
+            "type": "debezium-connector-operator",
+            "version": "[1.0.0,2.0.0)"
+          }
+        ],
+        "connector_type": "source",
+        "connector_class": "io.debezium.connector.postgresql.PostgresConnector",
+        "container_image": "quay.io/rhoas/cos-connector-debezium-postgres@sha256:81b1354c9a52b604715507952929a794f46d391807437bc0bd4d31eaeeba9d2b"
+      }
+    }
+  }
+}
